+++ ui/apps/platform/src/services/ReportsService.ts
@@ -68,7 +68,6 @@ export function fetchReportById(reportId: string): Promise<ReportConfiguration> 
     return axios
         .get<{ reportConfig: ReportConfiguration }>(`${reportConfigurationsUrl}/${reportId}`)
         .then((response) => {
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-return
             return response?.data?.reportConfig;
         });
 }
